\name{aslm-package}
\alias{aslm-package}
\alias{AdaptiveSparsity}
\alias{asggm-package}
\docType{package}
\title{
Adaptive Sparsity Models Model
}
\description{
implements the adaptive sparse linear model using Figueiredo's EM algorithm 
for adaptive sparsity (Jeffreys prior) and the adaptively sparse gaussian graphical model using Wong's parameter-free algorithm.
}
\author{
 Kristen Zygmunt, Eleanor Wong, Tom Fletcher

Maintainer: Kristen Zygmunt <krismz@sci.utah.edu>
}
\references{
\cite{Figueiredo, M.A.T.; , \dQuote{Adaptive sparseness for supervised learning}, Pattern Analysis and Machine Intelligence, IEEE Transactions on , vol.25, no.9, pp. 1150- 1159, Sept. 2003}

\cite{Wong, Eleanor, Suyash Awate, and P. Thomas Fletcher. \dQuote{Adaptive Sparsity in Gaussian Graphical Models.}In Proceedings of the 30th International Conference on Machine Learning (ICML-13), pp. 311-319. 2013.}
}
\keyword{ package }
\keyword{ models }
\keyword{ multivariate }
\seealso{
\code{\link{aslm}}, \code{\link{asggm}}
}
