\name{summary.aslm}
\alias{summary.aslm}
\alias{print.aslm}
\alias{print.summary.aslm}
\title{
        Handling aslm objects
}
\description{
summary and print methods for class \dQuote{\code{aslm}}
}
\usage{
\method{summary}{aslm}(object, \dots)
\method{print}{summary.aslm}(x, \dots)
\method{print}{aslm}(x, \dots)
}
\arguments{
  \item{object}{
An object of class \dQuote{\code{aslm}}, usually a result of a call to \code{\link{aslm}}
}
  \item{x}{
An object of class \dQuote{\code{summary.aslm}} or \dQuote{\code{aslm}}
}
  \item{\dots}{
Further arguments
}
}
\details{
summary and print methods to help display and work with aslm objects.
}
\value{
\code{print} prints a brief overview 
}
\author{
 Kristen Zygmunt, Eleanor Wong, Tom Fletcher
}

\seealso{
\code{\link{aslm}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
\keyword{ print }
